# Conflicts:
#	http/user.http
#	src/main/java/com/sparta/itsmine/domain/qna/controller/QnaController.java
#	src/test/java/com/sparta/itsmine/domain/qna/entity/QnaTest.java

+++ src/main/java/com/sparta/itsmine/domain/auction/dto/GetAuctionByUserResponseDto.java
@@ -2,6 +2,8 @@ package com.sparta.itsmine.domain.auction.dto;
 
 import com.querydsl.core.annotations.QueryProjection;
 import com.sparta.itsmine.domain.auction.entity.Auction;
+import com.sparta.itsmine.domain.product.entity.Product;
+import com.sparta.itsmine.domain.product.utils.ProductStatus;
 import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.Setter;
@@ -12,10 +14,10 @@ public class GetAuctionByUserResponseDto {
 
     private Long userId;
     private Long productId;
-    private Long bidPrice;
+    private Integer bidPrice;
 
     @QueryProjection
-    public GetAuctionByUserResponseDto(Long productId, Long bidPrice, Long userId) {
+    public GetAuctionByUserResponseDto(Long productId, Integer bidPrice, Long userId) {
         this.productId = productId;
         this.bidPrice = bidPrice;
         this.userId = userId;
